Add files via upload
Most of these files are directly borrowed from https://github.com/zsyOAOA/ResShift and are required for the running of the model, the files have been heavily edited by myself to enable the input of single-channel MRI, the use of a MobileViT Transformer setup and edited loss function. Smaller quality of life changes have also been made to enable easier tracking, visualisation and reviewing of data.